+++ src/app/app.component.ts
@@ -1,16 +1,15 @@
-import { AfterViewChecked, Component, ElementRef, ViewChild } from '@angular/core';
+import { AfterViewInit, Component, ElementRef, ViewChild } from '@angular/core';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.sass']
 })
-export class AppComponent implements AfterViewChecked {
+export class AppComponent implements AfterViewInit {
   @ViewChild('appTitle')
   appTitle!: ElementRef<HTMLHeadingElement>;
 
-  // Этот метод запускается периодически, если надо разово, то можно использовать ngAfterViewInit
-  ngAfterViewChecked(): void {
+  ngAfterViewInit(): void {
     this.appTitle.nativeElement.innerHTML = "Anguler приложение для покупки товаров";
   }
 }